+++ serialport/src/main/java/com/inno/serialport/function/driver/RS485Driver.kt
@@ -186,7 +186,7 @@ class RS485Driver : IDriver {
         val command = ((buffer[FRAME_CMD_INDEX_LOW].toInt() and 0xFF) shl 8) or
                 (buffer[FRAME_CMD_INDEX_HIGH].toInt() and 0xFF)
         val content = buffer.sliceArray(FRAME_CONTENT_START_INDEX until buffer.size - 2)
-        return PullBufInfo(length, command, content)
+        return PullBufInfo(length, command.toShort(), content)
     }
 
     private fun validPullInfo(buffer: ByteArray): PullBufInfo? {
